+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.6] - 2016-08-15
+### Fixed
+- Compile under the *c90* standard for better future portability.
+- Don't overwrite more useful error messages with the useless "The statement
+has been terminated." error.
+
+### Added
+- Minor documentation improvements.
+
 ## [1.0.5] - 2016-05-23
 ### Fixed
 - Fix the `ctds.Connection.timeout` property to raise an appropriate error